+++ crates/maple_cli/src/cmd/rpc/mod.rs
@@ -1,4 +1,6 @@
 mod filer;
+mod on_move;
+mod types;
 
 use std::io::prelude::*;
 use std::thread;
@@ -47,8 +49,12 @@ fn loop_handle_message(rx: &crossbeam_channel::Receiver<String>) {
         thread::spawn(move || {
             // Ignore the invalid message.
             if let Ok(msg) = serde_json::from_str::<Message>(&msg.trim()) {
+                log::debug!("recv msg: {:?}", msg);
                 match &msg.method[..] {
                     REQUEST_FILER => filer::handle_message(msg),
+                    "client.on_move" => {
+                        let _ = on_move::handle_message_on_move(msg);
+                    }
                     _ => write_response(json!({ "error": "unknown method", "id": msg.id })),
                 }
             }