+++ CHANGELOG.md
@@ -3,9 +3,14 @@ CHANGELOG
 
 ## [unreleased]
 
+### Added
+
+- Async `on_move` impl for `filer`, `files`, `grep` and `grep2` provider in Rust binary, no delay for the preview function. ([#437](https://github.com/liuchengxu/vim-clap/pull/437))
+
 ### Changed
 
 - Decrease the max number of candidates for running in sync from 100000 to 30000, which means once the total number of candidates is larger than 30000, the async filter will be used, otherwise use the builtin sync one.
+- `filer` uses the daemon job which requires the latest binary. Download the latest binary if you uses the prebuilt binary.
 
 ### Improved